feature(generator):
1. COGenerator inherit from COCoroutine
2. Unify yield and yield_val
3. separate cocore

+++ coobjc/promise/COPromise.h
@@ -210,7 +210,7 @@ typedef void (^COPromiseConstructor)(COPromiseFullfill fullfill, COPromiseReject
 - (void)setupWithProgress:(NSProgress*)progress;
 
 //get the next progress fractionCompleted value, this method should be called in a coroutine
-- (CGFloat)next;
+- (float)next;
 
 @end
